+++ gestion-stocks-middle/src/main/java/net/blog/dev/gestion/stocks/middle/api/IDetailStockMService.java
@@ -30,5 +30,5 @@ public interface IDetailStockMService {
 	 * @param stockListBean
 	 * @param quantity
 	 */
-	void updateQuantity(StockListBean stockListBean, Integer quantity);
+	void updateQuantity(StockListBean stockListBean, Float quantity);
 }
